change webworker to single file to reduce overall build size
